move log config up to getcall
we have WithLogger already, we just need to use it. this moves the config of
the logger itself up to be a call option which makes it less brittle in
GetCall to hit the path where it might get accidentally turned on.

removes the old buffer and io.Reader on stderr so that we could upload logs,
we only need the logrus logger for now and it was a bit of a tangled mess.

TODO need to add config bit to set the default level to 'info' so that we get
logs out of these guys in the agent config (pure runner is left alone to off),
with ability to change to debug or turn off altogether.

NOTE: this fixes #1328 by putting a write guard after close and not using the
line writer for different writers